+++ CHANGES.rst
@@ -10,7 +10,9 @@ Date: (TBD)
 - remove udisks1 support (and corresponding command line options)
 - use asyncio+gbulb internally
 - internal refactoring and simplifications
-- keep password dialogs on top
+- improve password dialog:
+  - "show password" checkbox
+  - keep password on top
 
 1.7.2
 ~~~~~